controls: s/^cdef void/cdef/
It appears that declaring a function "cdef void" can inhibit exception
propagation. I only observed this in effect on inverted tag blocks even
though the code structure was much the same for classic tag blocks.
Online guidance suggests changing it to a vanilla "cdef" to denote an
object return type and to enable exception propagation correctly.