controls: Add block aliases
`classic-tag-block` and `inverted-tag-block` aren't as communicative as
they could be. This change introduces new aliases for these types of
blocks - `span-tag-block` and `tag-span-block`, respectively, which
spell out their internal hierarchies exactly.

The old names (`classic-` and `inverted-`) continue to be supported for
backwards compatibility, but I consider the new names canonical.